Add paste crate for testing
This allows concatenation of ident values when defining
a macro. As the concat_ident macro in std is not likely
to ever make it into stable, this is the best bet.

Using this means that test function prefixes can be templated
out and only the suffixes need be passed in.