utils: add helper to generate sample configs.
In an effort to make the examples easier to maintain, this is a one-off
tool that will take a config in one format, and output the other two
formats. This is also handy if you can't quite visualize a configuration
in a particular format (e.g. toml) but can in another (e.g. json).